feat: push album art files to s3

+++ src/services/maps/s3_handler.ts
@@ -1,6 +1,11 @@
-import { DeleteObjectCommand, PutObjectCommand, S3Client } from '@aws-sdk/client-s3';
+import {
+  DeleteObjectsCommand,
+  ListObjectsV2Command,
+  PutObjectCommand,
+  S3Client,
+} from '@aws-sdk/client-s3';
 import { checkExists } from 'base/preconditions';
-import { Result } from 'base/result';
+import { PromisedResult, Result } from 'base/result';
 import * as fs from 'fs/promises';
 import * as path from 'path';
 import { getEnvVars } from 'services/env';
@@ -33,23 +38,25 @@ function getS3Client() {
   return s3;
 }
 
-export async function uploadFiles(opts: {
-  id: string;
-  mapsDir: string;
-  buffer: Buffer;
-  albumArtFiles: unzipper.File[];
-}): Promise<Result<string | undefined, S3Error>> {
-  // Upload it to the bucket
+async function s3Put(
+  key: string,
+  buffer: Buffer,
+  contentType: string
+): PromisedResult<undefined, S3Error> {
   try {
     const s3 = getS3Client();
     await s3.client.send(
       new PutObjectCommand({
         Bucket: s3.bucket,
-        Key: `maps/${opts.id}.zip`,
-        Body: opts.buffer,
-        ContentType: 'application/zip',
+        Key: key,
+        Body: buffer,
+        ContentType: contentType,
       })
     );
+    return {
+      success: true,
+      value: undefined,
+    };
   } catch (e) {
     return {
       success: false,
@@ -62,24 +69,59 @@ export async function uploadFiles(opts: {
       ],
     };
   }
+}
 
-  // TODO: write it to disk and upload it to the bucket via a worker thread to speed up the
-  // upload map flow.
-  // Then, on download, check whether it's on disk and serve it from there first.
-  // When the upload completes, delete it from disk.
-  // const mapFilename = opts.id + '.zip';
-  // const filepath = path.resolve(opts.mapsDir, mapFilename);
-  // await fs.writeFile(filepath, opts.buffer);
+async function s3Delete(keys: string[]): PromisedResult<undefined, S3Error> {
+  try {
+    const s3 = getS3Client();
+    await s3.client.send(
+      new DeleteObjectsCommand({
+        Bucket: s3.bucket,
+        Delete: {
+          Objects: keys.map((k) => ({ Key: k })),
+        },
+      })
+    );
+    return { success: true, value: undefined };
+  } catch (e) {
+    return { success: false, errors: [{ type: S3Error.S3_DELETE_ERROR }] };
+  }
+}
 
-  // For now, write all of the album art files to the on-disk maps directory.
+function guessContentType(filename: string): string {
+  if (filename.endsWith('png')) {
+    return 'image/png';
+  } else if (filename.endsWith('jpg') || filename.endsWith('jpeg')) {
+    return 'image/jpeg';
+  } else if (filename.endsWith('bmp')) {
+    return 'image/bmp';
+  } else if (filename.endsWith('gif')) {
+    return 'image/gif';
+  } else if (filename.endsWith('webp')) {
+    return 'image/webp';
+  }
+  return 'application/octet-stream';
+}
+
+export async function uploadFiles(opts: {
+  id: string;
+  buffer: Buffer;
+  albumArtFiles: unzipper.File[];
+}): Promise<Result<string | undefined, S3Error>> {
+  // Upload map file to the bucket
+  const mapPutResult = await s3Put(`maps/${opts.id}.zip`, opts.buffer, 'application/zip');
+  if (!mapPutResult.success) {
+    return mapPutResult;
+  }
+
+  // Write album art files to S3
   // TODO: display all of the album arts in the FE, e.g. in a carousel, or when selecting a difficulty
-  const albumArtFolderPath = path.resolve(opts.mapsDir, opts.id);
-  await fs.mkdir(albumArtFolderPath, { recursive: true });
   await Promise.all(
-    opts.albumArtFiles.map((a) => {
+    opts.albumArtFiles.map(async (a) => {
       const albumArt = checkExists(a, 'albumArt');
-      const albumArtPath = path.resolve(albumArtFolderPath, path.basename(albumArt.path));
-      return albumArt.buffer().then((b) => fs.writeFile(albumArtPath, b));
+      const buffer = await albumArt.buffer();
+      const filename = path.basename(albumArt.path);
+      return s3Put(`albumArt/${opts.id}/${filename}`, buffer, guessContentType(filename));
     })
   );
 
@@ -90,21 +132,31 @@ export async function uploadFiles(opts: {
   };
 }
 
-export async function deleteFiles(opts: {
-  mapsDir: string;
-  id: string;
-}): Promise<Result<undefined, S3Error>> {
-  try {
-    const s3 = getS3Client();
-    await s3.client.send(
-      new DeleteObjectCommand({ Bucket: s3.bucket, Key: `maps/${opts.id}.zip` })
-    );
-  } catch (e) {
-    return { success: false, errors: [{ type: S3Error.S3_DELETE_ERROR }] };
+export async function deleteFiles(opts: { id: string }): Promise<Result<undefined, S3Error>> {
+  const mapDeleteResult = await s3Delete([`maps/${opts.id}.zip`]);
+  if (!mapDeleteResult.success) {
+    return mapDeleteResult;
   }
 
-  const albumArtFolderPath = path.resolve(opts.mapsDir, opts.id);
+  const { mapsDir } = getEnvVars();
+
+  // Attempt to remove from both disk and S3
+  const albumArtFolderPath = path.resolve(mapsDir, opts.id);
   await fs.rm(albumArtFolderPath, { recursive: true, force: true });
+  try {
+    const s3 = getS3Client();
+    const albumArtFiles = await s3.client.send(
+      new ListObjectsV2Command({
+        Bucket: s3.bucket,
+        Prefix: `albumArt/${opts.id}`,
+      })
+    );
+    await s3Delete(
+      albumArtFiles.Contents?.map((c) => c.Key).filter((k): k is string => k != null) || []
+    );
+  } catch (e) {
+    // TODO: log error to Sentry
+  }
 
   return { success: true, value: undefined };
 }
